Show contract address loader only on LocalAnchor
