+++ app/src/@anchor-protocol/app-provider/env.ts
@@ -88,5 +88,5 @@ export enum ANCHOR_QUERY_KEY {
   EVM_BLOCK_NUMBER = 'EVM_BLOCK_NUMBER',
   PROJECTED_EARN_APY = 'PROJECTED_EARN_APY',
   BLOCK = 'BLOCK',
-  ANCHOR_CONTRACT_ADDRESSES = 'ANCHOR_CONTRACT_ADDRESSES',
+  LOCAL_ANCHOR_CONTRACT_ADDRESS = 'LOCAL_ANCHOR_CONTRACT_ADDRESS',
 }
