+++ app/src/queries/index.ts
@@ -11,3 +11,4 @@ export * from './evm/useEvmTerraAddressQuery';
 export * from './evm/useEvmBlockNumberQuery';
 export * from './useAnchorQuery';
 export * from './astroport/useAssetPriceInUstQuery';
+export * from './useLocalAnchorContractAddressQuery';
